+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Change Log
 
+## 0.0.8
+- Added "Add cs-icon class" CodeLens to SVGs which do not have it
+
 ## 0.0.7
 - Added navigator for HTML/nunjucks files
 
 ## 0.0.6
-- Added functionality to make Nav list 11ty compliant
+- Added functionality to make Nav list 11ty compliant with CodeLens